+++ src/components/Otsu.vue
@@ -12,186 +12,57 @@
 -->
 
 <template>
-  <div class="Otsu">
-    <h1>{{ msg }}</h1>
-    <v-container class="text-xs-left">
-
-      <v-layout row wrap>
-        <v-flex xs12>
-          <h3>Description</h3>
-          <p>Upload an image or ZIP file containing set of images (Supported file formats: .jpg, .tif, .png) and click
-            "Binarize" to binarize image using Otsu's Method.</p>
-        </v-flex>
-        <v-flex xs12>
-          <h3>Input Options</h3>
-          <p><strong>1. Single image:</strong>Supported image formats are .jpg, .tif and .png.
-          <p><strong>2. Single image in .mat format:</strong> The .mat file must contain ONLY ONE variable named "Input," which contains the image.</p>
-          <p><strong>3. ZIP file with multiple images:</strong> Submit a ZIP file containing multiple images (supported formats: .jpg, .tif, .png) of same size (in pixels). DO NOT ZIP the folder containing images; select all images and ZIP them directly.</p>
-        </v-flex>
-      </v-layout>
-
-      <v-alert
-        v-model="loginRequired"
-        type="error"
-        outline
-      >
-        {{loginRequiredMsg}}
-      </v-alert>
-
-      <v-alert
-        v-model="errorAlert"
-        type="error"
-        dismissible
-      >
-        {{errorAlertMsg}}
-      </v-alert>
-      
-      <v-dialog v-model="successDlg" persistent max-width="500px">
-        <v-card>
-          <v-card-title>
-            <span>Otsu Binarization Job Submitted Successfully</span>
-            <v-spacer></v-spacer>
-          </v-card-title>
-          <v-card-text>
-            Your Otsu job is: {{jobId}} <br/> You should receive an email with a link to the job output.
-          </v-card-text>
-          <v-card-actions>
-            <v-btn color="primary" flat @click="successDlgClicked()">Close</v-btn>
-          </v-card-actions>
-        </v-card>
-      </v-dialog>
-
-      <ImageUpload v-on:setFiles="setFiles" aspectRatio='free'></ImageUpload>
-
-      <v-flex class="text-xs-center">
-        <v-btn id="binarize-button" v-on:click="submit()" color="primary">Binarize</v-btn>
-      </v-flex>
-
-      <v-flex xs12>
-        <h3>References</h3>
-        <p>N. Otsu, A threshold selection method from gray-level histograms, IEEE transactions on systems, man, and
-          cybernetics, vol. 9, no. 1, pp. 62-66, 1979.</p>
-      </v-flex>
-
-    </v-container>
-  </div>
+  <McrJobsTemplate :job='jobInfo'></McrJobsTemplate>
 </template>
 
 <script>
-import {} from 'vuex'
-import {JobMgr} from '@/modules/JobMgr.js'
-import {Auth} from '@/modules/Auth.js'
-import ImageUpload from './ImageUpload.vue'
 
-export default {
-  name: 'Otsu',
-  components: {
-    ImageUpload
-  },
-  data: () => {
-    return ({
-      title: 'Input Upload',
-      msg: 'Binarization - Otsu Method',
-      dialog: false,
-      errorAlert: false,
-      errorAlertMsg: '',
-      loginRequired: false,
-      loginRequiredMsg: '',
-      successDlg: false,
-      jobId: '',
-      files: [],
-      fileName: ''
-    })
-  },
+  import {} from 'vuex'
 
-  beforeMount: function () {
-    let vm = this
-    vm.auth = new Auth()
-    if (!vm.auth.isLoggedIn()) {
-      vm.loginRequired = true
-      vm.loginRequiredMsg = 'Login is required.'
-    }
-  },
+  export default {
 
-  methods: {
+    name: 'Otsu',
 
-    setFiles: function (...files) {
-      this.files = files[0]; // the actual file object
-      this.fileName = files[1]; // the name of the file
+    components: {
+      McrJobsTemplate
     },
 
-    setLoading: function () {
-      this.$store.commit('isLoading')
-    },
+    data: () => {
+      return {
+        jobInfo: {
 
-    resetLoading: function () {
-      this.$store.commit('notLoading')
-    },
+          jobTitle: 'Otsu',
 
-    successDlgClicked: function () {
-      let vm = this
-      console.log('Success dlg button clicked')
-      vm.$router.go(-2) // go back to mcr homepage page
-    },
+          pageTitle: 'Binarization - Otsu Method',
 
-    submit: function () {
-      let vm = this
-      vm.files.forEach(function (v) {
-        console.log(JSON.stringify(v))
-      })
+          description: 'Upload an image or ZIP file containing set of images (Supported file formats: .jpg, .tif, .png) and click "Binarize" to binarize image using Otsu\'s Method.',
+          
+          aspectRatio: 'free',
+          
+          submitButtonTitle: 'Binarize',
+          
+          uploadOptions: [
+            {
+              title: 'Single image', 
+              description: 'Supported image formats are .jpg, .tif and .png.'
+            },
+            {
+              title: 'Single image in .mat format',
+              description: 'The .mat file must contain ONLY ONE variable named "Input," which contains the image.'
+            },
+            {
+              title: 'ZIP file with multiple images',
+              description: 'Submit a ZIP file containing multiple images (supported formats: .jpg, .tif, .png) of same size (in pixels). DO NOT ZIP the folder containing images; select all images and ZIP them directly.'
+            }
+          ],
 
-      vm.setLoading()
-      console.log('Loading..')
-      let jm = new JobMgr()
-      console.log('Called Job Manager')
-      jm.setJobType('otsu')
-      jm.setJobParameters({'InputType': vm.fileName.split('.').pop()}) // Figure out which input type
-      if (vm.files && vm.files.length >= 1) {
-        vm.files.forEach(function (v) {
-          jm.addInputFile(v.fileName, v.fileUrl)
-          console.log('Job Manager added file: ' + v.fileName)
-        })
-        return jm.submitJob(function (jobId) {
-          console.log('Success! JobId is: ' + jobId)
-          vm.jobId = jobId
-          vm.resetLoading()
-          vm.successDlg = true
-        }, function (errCode, errMsg) {
-          let msg = 'error: ' + errCode + ' msg: ' + errMsg
-          console.log(msg)
-          vm.errorAlertMsg = msg
-          vm.errorAlert = true
-          vm.resetLoading()
-        })
-      } else {
-        let msg = 'Please select a file to process.'
-        vm.errorAlertMsg = msg
-        vm.errorAlert = true
-        vm.resetLoading()
+          references: [
+            'N. Otsu, A threshold selection method from gray-level histograms, IEEE transactions on systems, man, and cybernetics, vol. 9, no. 1, pp. 62-66, 1979.'
+          ]
+
+        }
       }
     }
   }
-}
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-  img {
-    width: 240px;
-  }
-
-  h1 {
-    margin-top: 10px;
-    background-color: black;
-    color: white;
-  }
-
-  h3 {
-    margin-bottom: 15px;
-  }
-  
-  p {
-    margin-left: 15px;
-  }
-
-</style>
